validation関連の整理
